Use Tex Live 2024 (#65)
Minted 3.7.0 introduced a bug for (certain instances) of mathescape,
namely that it looped forever on code/append_annot.py.
https://github.com/gpoore/minted/issues/451

Even though version 3.7.0 came into development on the 8th of April:
https://github.com/gpoore/minted/commit/7d7a2694dd07406a57dac8ce7be97b1571dc9ba2
and was officially released on 15th of May:
https://github.com/gpoore/minted/commit/19ea6fcae5a6360cbef65dd4443df7cae05a5e49
and the TeX Live 2025 was release on 8th March, 2025, putting "2025"
as the version in the setup-texlive-action still gave minted version
3.7.0.

Hence this commit now pins the Tex Live version to 2024.

PS. Minted as a debug=true option that was helpful in figuring out which
files were causing the issue. However, it creates a *.config.minted
file and updates it on each run when using latexmk, which detects the
change and aborts execution after the maximum run limit. Hence debug
cannot be left on by default.