Fix #51 and #52
This commit adds a pseudo-ResCore example and informally walks through the type checking process for it. In particular, its intention is twofold. First, to show an example of a CN spec turned into a Kernel CN type. Second, to clarify how resources from the types of memory actions end up in the context via pattern-matching in let-binders.