Start explaining proof and soundness
The rules need to be tweaked a bit. Note that

|       | pointer | rt/integer | integer |
|-------+---------+------------+---------+
| CN    | NULL    | (pi,0)     | 0       |
| VIP   | NULL    | Int (0)    | Int(0)  |
| CN    | (pi, a) | (pi,a)     | a       |
| VIP   | (pi, a) | Loc (pi,a) | Int (a) |

and that the typing rule for rt to pointer needs to provide an
unconstrained provenance if it cannot prove the allocation is live
(doesn't mean it's not live, just means it can't prove it).

Need to update the inference, dynamics and proof and also add in the new
rule.